+++ deepinv/physics/__init__.py
@@ -26,12 +26,4 @@ from .remote_sensing import Pansharpen
 
 from . import functional
 from . import generator
-
-# from .generator import (
-#     PhysicsGenerator,
-#     GeneratorMixture,
-#     MotionBlurGenerator,
-#     DiffractionBlurGenerator,
-#     AccelerationMaskGenerator,
-#     SigmaGenerator
-# )
+from .phase_retrieval import PhaseRetrieval, RandomPhaseRetrieval